% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_create_excel.R
\name{ds_create_excel}
\alias{ds_create_excel}
\title{Create Excel Sheets for Semi-Automated Datastream Download}
\usage{
ds_create_excel(
  .data,
  .col_index,
  .formula,
  .dir,
  .name,
  .split = 1000,
  .progress = TRUE
)
}
\arguments{
\item{.data}{A Dataframe with company Identifier used by Datastream\cr
Notes:\cr
\itemize{
\item The names of the columns do not matter.\cr
\item The Dataframe can contain any number of additional columns
}}

\item{.col_index}{The Index of the column that contain Datstream Identifiers (1,2,3,...)}

\item{.formula}{The whole formula Datastream uses to get your query, the first arguments MUST be exchanged by "{range}" and the formula MUST be enquoted in single-quotes\cr
Example:\cr
'=DSGRID({range};"VO;UVO;RI";"2004-01-01";"2009-12-31";"D";"RowHeader=true;ColHeader=true;Transpose=true;Code=true;DispSeriesDescription=true;YearlyTSFormat=false;QuarterlyTSFormat=false")'}

\item{.dir}{Full Path of the folder where the excel sheets should be stored}

\item{.name}{Prefix for the Excel Files}

\item{.split}{Numbe rof Companies in one Excel Sheet}

\item{.progress}{Show Progress?}
}
\value{
Excel Sheets
}
\description{
Create Excel Sheets for Semi-Automated Datastream Download
}
